backend/requirements.txt
# Radar Simulation Platform - Backend Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# WebSocket
python-socketio>=5.10.0
aiohttp>=3.9.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.25.0

# CORS (FastAPI has built-in CORS support, no extra package needed)
# python-cors is not a valid package, removed

# Rate Limiting
slowapi>=0.1.9

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0

# Data Storage
h5py>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
faker>=25.0.0
freezegun>=1.4.0

# Development
black>=23.12.0
flake8>=6.1.0
